[3.2] Locate unique claim by scanning canvas
The unique claim would have all its pixels with value 1.  Strictly
just checking the stroke border is enough but that would make the code
complex, so going with a fill check.

Visualization of the claims was done using minifb when investigating
the benefit of optimising with a bounding volume hierarhy (BVH) like
AABB tree.  However, since L3 caches these days are usually in the
order of MiBs, the entire canvas vector of 1000 × 1000 u8s (≅ 1 MiB)
should fit, not warranting the need for a spatial data structure.