[4.1] Find most commonly asleep minute
The problem is essentially finding the most overlapping point in a 1D
array of intervals that can be as long as 60 elements.

Using an array of 60 u8s seems like a better solution here too.  The
entire memory would be in stack (240 bytes even if each u8 was made
into u32), fairly small, fast and leads to simple code.

Alternatives are around sorting and finding intersections, but the
complexity can become exorbitant.  Segment trees are promising but
tree data structures with pointers are usually slower than arrays for
small workloads.

https://www.geeksforgeeks.org/check-if-any-two-intervals-overlap-among-a-given-set-of-intervals/