[4.2] Find guard with highest asleep frequency/min
Since finding the most frequent minute of sleepiness is common to both
strategies move it to a separate function.  Its pair return type seems
arcane to the uninformed, make a tuple struct for better readability.

Strategy 2 is fairly simple, at least in code.  It’s just finding the
max element in a collection with a computed key.