[5.2] Find shortest polymer length by filtering
Before folding the polymer, filtering the incoming character for all
chars in A..=Z is the obvious choice; it’s an O(m.n) solution, where m
in the input set, n is the letters in the alphabet.

Have input iteration in the outer loop and letter iteration in the
inner loop to save from storing the input; reversing the loops would
mean wasting 50000 bytes.