[6.1] Brute-force solution
* For every cell in grid
  + Calculate taxicab distance to all locations
  + Sort candidates based on distance
  + If not a tie, increment location ID’s frequency
  + If in edge, mark location ID’s frequency as ∞
* Print location with maximum frequency that’s not ∞

Before choosing to go brute-force, as an optimisation I tested only
the surrounding locations of a cell instead of every location.
However, this approach isn’t correct; gave wrong results.  Example, in
the below case, we’d test c but not o, the closest location.

       . c
       .
       .
       .
       X...o

Either due to misunderstanding on my part or poor example on the
puzzle’s part led to the impression that only the locations that are
in the grid’s boundary can extend to infinity.  However, even a
location not in the border can extend to infinity; see floodfill.
Essentially we’re calculating a voronoi diagram using taxicab distance.

Calculating the location frequency in the grid taking above into
consideration gives the correct answer.