[7.1] Dependency graph using array and Vec
Simple directed acyclic graph in Rust using an array and Vec<u8>.
Decided against using a hash map as an array with 26 elements will do
as we infer from the input that tasks ∈ [A, Z].

Instead of storing the dependencies of a task, stored the depenedents,
so that when a task is over it can go through all its dependants and
notify -- push, instead of pull, model.

Linearize tasks by completing ones with least dependency and tie
breaking lexicographically.