[7.2] Move task end time to thread context
Since end time needn’t be associated to waiting tasks it doesn’t make
sense to put this data unnecessarily for all tasks. It’s needed only
for the tasks that are run by threads, so it makes sense to keep this
information along with the thread, the doer of the task.