[10] Deduce message with time taken
The conditioned graphics programmer in me chose the route of plotting
the stars on a canvas instead of printing it on a terminal - a much
simpler option.  I could get the solution working after battling with
off-by-one errors not showing the pixels within the canvas.  The
solution still requires manual intervention of pausing the simulation
and proceeding framewise when the message is imminent.  This could’ve
been avoided with the following approaches.

A key observation I missed but found in another solution [1] to this
puzzle was that the bounding box shrinks on every time step and
reaches minimum when the message is formed and grows from thereon.

Another solution [2] instead used a interesting approach: time step
silently as long as bounding box is > 80x80 and print when it’s
within. It printed some 4 iteration worth sky plots out of which one
was the actual message.

Lots to learn!

[1]: https://github.com/KanegaeGabriel/advent-of-code-2018
[2]: https://github.com/BurntSushi/advent-of-code