[11.2] Generalize finding maximum-powered subarray
Experimented with using Kadane’s algorithm in 2D and wasted quite a
bit of time.  Realized that it doesn’t scale up straight-forwardly.

Got out of the quagmire of analysis-paralysis after seeing [1] and
realizing that the brute force approach is perhaps slow but the
development time is much faster.  Though this isn’t a full brute
force implementation but a generalization of the same cache-friendly
one of Part 1’s.  It’s also close to the optimal solution [2] using
summed area tables [3].

Intel i7-8850H (12) @ 2.6 GHz (MacBook Pro 15’ 2018) solves both parts
in 1.796 seconds.

[1]: https://github.com/BurntSushi/advent-of-code/
[2]: https://github.com/KanegaeGabriel/advent-of-code-2018/
[3]: https://en.wikipedia.org/wiki/Summed-area_table