[15.1] Exit condition and outcome of battle
* Use ‘while { }’ construct to simulate battle but also have label to
  allow an inner loop to break out.
* Use a hash map instead of vector to store fighters against their ID
  as it’s easier to lookup by ID despite sorting of fighters every round
* Mark but not remove dead fighters to avoid removing their IDs from Vec
  sorted for turn
* Implement Map::attack to deal damage to a unit by ID
* Fix Figher::target() to return enemy ID after sorting
  by (hits, position)
* Use unreachable!() instead of panic!() where it makes sense
* Use iter::{copied,cloned} instead of map (thanks to clippy)

As usual the puzzle description was quite confusing and had lot of
nuances that could’ve been better explained.  Two gotchas:

1. A fighter just moving into position for an attack can also attack
   in the same turn
2. Can’t stop the battle when the last enemy’s fallen

Only when the next fighter notices it can the battle be ended, without
counting the noticing round.  The subtlety is that the next fighter
may or may not be in the same round; hence the use of both while and
break.

The many sample test boards was helpful though.