[19.2] Optimize brute force elfcode with Rust
Part 2 of running the same elfcode with accumulator set to 1 led to a
really long wait.  Part 1 was too easy, so I should’ve expected it ;)

Had to understand that the assembly code is trying to resolve a large
number into its factors and accumulate its sum in accumulator (reg 0).

This didn’t immediately give the solution since I only had prime
factorization code handy but had to understand how to compute all
factors of a number from its prime factors.  Implemented this
factorization in Rust; along the way learned to implement

1. Cartesean product for multiple lists
2. Rust modules