[20] Add puzzle description
Charting out the entire map and then doing a BFS is needless; a much
simpler solution is to follow the path described in the input
expression and update the room distances as we go [1]; this is a
single pass solution that doesn’t need allocation or path finding!  I
suspected something like this earlier but my conditioning of earlier
solved AoC puzzles’ complexity dissuaded me from this.

[1]: https://github.com/KanegaeGabriel/advent-of-code-2018/blob/master/20_a_regular_map.py