[24.2] Compute alive immune units after minimal boost
Due to me missing a rule in the puzzle led to a long bug chasing
session and was able to find the issue only afer pour over [1].  The
rule was in not choosing an enemy group if no damage can be dealt.

`fight()` was stuck in an infinite loop.  Thanks to [2] which made me
realize there’s a possibility of a battle never-ending which can be
broken out of by checking if there’re any units lost.

* Introduce `boost: u16` and demote `damages` to `u16` to maintain
  `Group` as a 4 x 4 byte structure.
* Keep quantities unsigned as much as possible and cast to signed
  integer only at sorting points.
* Introduce a `Victor` struct to hold the results of a fight.
* In a fight, to filter out active combats, instead of `enumerate()`
  to walk `ids` and `choices` in lock-step, `zip()` is the more
  appropriate option.
* Fuse `filter` and `filter_map` into the latter.
* Solved part 2 using linear search.

[1]: https://todd.ginsberg.com/post/advent-of-code/2018/day24/
[2]: https://github.com/KanegaeGabriel/advent-of-code-2018