[2015.5.2] [Python] Count really nice strings
Missed seeing the "aaaa" case for quite a while; it tripped a
simple (counter) dict lookup; once realized, quickly solved it by
keeping a list of occurances solved it.

Fixed Part 1 code in not checking all pairs of a string e.g. in a
"abcd" the code checked only "ab" and "cd", but not "bc".  Wrote a
sliding window generator (most satisfying part of the exercise
for me) for this.