[2015.10.2] [Python] RLE look-and-say
As iterations grow the elements in the series become exponentially
large; it didn’t occur to me that memory allocation would be the
bottleneck and run-length encoding would be great savings.  I cheated
and looked up some solutions online as I was side-tracked by the
red-herring that’s Conway’s video [1] as mentioned in the question.

RLE solution computed all 50 iterations in approximately 6.4 seconds
on the same machine mentioned in the parent commit.

Solution itself is simple; heavily used itertools.

[1]: https://www.youtube.com/watch?v=ea7lJkEhytA