[2025.7.2] Count the many worlds of Tachyon beam
A simple, straight-forward approach was a nested `for` over lines and
`vector<beams>` which start at one.  Split and add to vector’s end
which are iterated over from next iteration onwards.  This works for
the sample input but loops forever for the actual input.

After looking at the sample input for a while it occured to me that
there would be a lot of duplicate beams living the same life as its
counterparts.  Initially thought I’d keep a multiplier but it was a
no-go.  Then it occured to me that it’s almost a pascal triangle!
Keep count of the beam count in a cell by editing the manifold (now
made of integers) and it’s smooth sailing thereon :)

Happy with the result as I solved this on Christmas eve in a train
back home 😇.